refactor: post 도메인 오류 해결
