+++ src/main/java/com/pharmquest/pharmquest/domain/post/data/mapping/PostLike.java
@@ -1,8 +1,8 @@
 package com.pharmquest.pharmquest.domain.post.data.mapping;
 
 import com.pharmquest.pharmquest.domain.post.data.Post;
+import com.pharmquest.pharmquest.domain.user.data.User;
 import com.pharmquest.pharmquest.global.data.BaseEntity;
-import com.pharmquest.pharmquest.user.domain.User;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.DynamicInsert;
