--- src/main/resources/templates/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>인덱스 페이지</title>
-</head>
-<body>
-<h1>인덱스페이지입니다.</h1>
-</body>
-</html>
