--- src/main/resources/templates/loginForm.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>로그인 페이지</title>
-</head>
-<body>
-<h1>로그인 페이지</h1>
-<hr/>
-<form action="/login" method="POST">
-    <input type="text" name="username" placeholder="Username"/><br/>
-    <input type="password" name="password" placeholder="Password"/><br/>
-    <button>로그인</button>
-</form>
-<a href="/oauth2/authorization/google">구글 로그인</a></br>
-<a href="/oauth2/authorization/naver">네이버 로그인</a>
-<a href="/joinForm">회원가입을 아직 하지 않으셨나요?</a>
-</body>
-</html>
